+++ NEWS.md
@@ -1,5 +1,9 @@
 # glyparse (development version)
 
+## New features
+
+* Add `parse_glycam_iupac()` to parse GlyCAM IUPAC structures. (#25)
+
 ## Minor improvements and bug fixes
 
 * `parse_wurcs()` now supports additional generic WURCS residue descriptors. (#23)
@@ -136,9 +140,11 @@
 
 * Better error messages.
   Before:
+
   ```r
   > parse_iupac_condensed("bad_glycan")
   ```
+
   ```
   Error in `purrr::map()` at glyparse/R/struc-parser-wrapper.R:13:3:
   ℹ In index: 1.
@@ -147,10 +153,12 @@
   ℹ Invalid characters or format in IUPAC-condensed string
   Run `rlang::last_trace()` to see where the error occurred.
   ```
+
   Now:
   ```r
   > parse_iupac_condensed("bad_glycan")
   ```
+
   ```
   Error in `parse_iupac_condensed()`:
   ! Can't parse: "bad_glycan"
@@ -177,3 +185,4 @@
 * `parse_iupac_short()`, `parse_iupac_extended()`, `parse_iupac_condensed()`,
   `parse_wurcs()` now support multiple substituents on the same monosaccharide,
   to align with the updates in `glyrepr` v0.5.0.
+